Create Learning PostgreSQL
The objective of this file was to introduce myself to another SQL language, PostgreSQL.

I was given a database of two tables. One with info on people within Netflix titles and the other with further information on those shows and movies.

I was given 5 questions to answer:
How many movie titles are there in the database? (movies only, not tv shows)
When was the most recent batch of tv shows and/or movies added to the database?
List all the movies and tv shows in alphabetical order.
Who was the Director for the movie Bright Star?
What is the oldest movie in the database and what year was it made?